Update style.css
using the given quote svg instead of the pseudo element for the that icon over daniel card.
then updating the margins of the page to suit the width of the browser better by using the max-width property on the main element. 
then reducing the number of columns of the main element for displays under 1080p width by one column for the better.